modified the long line of codes and removed redundant codes
