+++ src/components/services/ServiceCard.jsx
@@ -1,4 +1,6 @@
+"use client";
 import { useState } from "react";
+
 import { motion } from "framer-motion";
 import { ArrowRight } from "lucide-react";
 import Card from "../common/Card";
